Extract JsonModelIo resource manifest data
Move JointedModelResource reflection into a package-private helper and add characterization coverage for reflected poses, joints, joint arrays, joint array ids, and root joints.
